+++ conanfile.py
@@ -10,7 +10,7 @@ required_conan_version = ">=2.0"
 
 class SISLConan(ConanFile):
     name = "sisl"
-    version = "14.6.1"
+    version = "14.7.0"
 
     homepage = "https://github.com/eBay/sisl"
     description = "Library for fast data structures, utilities"