+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [14.7.0] - 2026-06-04
+- Added new generic sisl::result generic type
+- Added new generic when_quorum async reactor
+
+## [14.6.1]
+
+## [14.6.0]
+
 ## [14.5.0] - 2026-06-04
 - Add deregister_generic_rpc()